+++ src/CustomFormatter.java
@@ -1,5 +1,3 @@
-package ButtonPanel;
-
 
 /**
  *
