+++ src/DisplayMessages.java
@@ -1,5 +1,3 @@
-package ButtonPanel;
-
 import imageconverter.ImageConverter;
 import java.io.File;
 import java.util.logging.Level;
