+++ src/MainDriver.java
@@ -1,5 +1,3 @@
-import ButtonPanel.ButtonPanel;
-import javax.swing.JFrame;
 import javax.swing.SwingUtilities;
 
 /*
@@ -13,13 +11,14 @@ import javax.swing.SwingUtilities;
  */
 public class MainDriver {
     
-    public static void createAndShowGUI() {
-        JFrame frame = new JFrame("Steganography");
-        frame.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
-        frame.setBounds(300, 300, 700, 400);
-        frame.setContentPane(new ButtonPanel());       
-        frame.setVisible(true);
-    }
+//    public static void createAndShowGUI() {
+//        JFrame frame = new JFrame("Steganography");
+//        frame.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
+//        frame.setBounds(100, 100, 400, 400);
+//        frame.add(new LeftPanel());
+//        frame.add(new RightPanel());
+//        frame.setVisible(true);
+//    }
     /**
      * main() is the initializer and executes the GUI on an EDT as opposed to
      * the main thread
@@ -28,7 +27,9 @@ public class MainDriver {
      */
     public static void main(String[] args) {
         SwingUtilities.invokeLater(() -> {
-            createAndShowGUI();
+            ButtonPanel execute = new ButtonPanel("Steganography");
+            execute.setVisible(true);
+//            createAndShowGUI();
         });
     }
 }
